+++ OASMan_ESP32/src/manifoldSaveData.h
@@ -79,9 +79,9 @@ headerDefineSaveFunc(internalReboot, bool);
 // values moved from the user defines file
 headerDefineSaveFunc(bagMaxPressure, uint8_t);
 headerDefineSaveFunc(blePasskey, uint32_t);         // 6 digits base 10
-headerDefineSaveFunc(systemShutoffTimeM, uint64_t); // may have to change
+headerDefineSaveFunc(systemShutoffTimeM, uint32_t); // may have to change
 headerDefineSaveFunc(compressorOnPSI, uint8_t);
 headerDefineSaveFunc(compressorOffPSI, uint8_t);
-headerDefineSaveFunc(pressureSensorMax, uint8_t);
+headerDefineSaveFunc(pressureSensorMax, uint16_t);
 
 #endif
